server/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bidict==0.23.1
click==8.1.7
colorama==0.4.6
fastapi==0.112.2
frozenlist==1.4.1
h11==0.14.0
idna==3.8
multidict==6.0.5
pydantic==2.8.2
pydantic_core==2.20.1
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
websockets==13.0
wsproto==1.2.0
yarl==1.9.4
PyJWT==2.8.0
